move median fn to file
